requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.1
django-allauth==0.42.0
django-ckeditor==6.0.0
django-crispy-forms==1.9.2
django-fs-trumbowyg==0.1.4
django-heroku==0.3.1
django-js-asset==1.2.2
django-nested-admin==3.3.2
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
psycopg2==2.7.5
pylibmc==1.6.1
python-binary-memcached==0.30.0
python-dotenv==0.14.0
python-monkey-business==1.0.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
uhashring==1.2
urllib3==1.25.10
whitenoise==5.2.0
